backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.16.2
openai==1.3.6
pydantic==2.5.0
pydantic-settings==2.1.0
transformers==4.45.0
torch==2.4.0
asyncpg==0.29.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
markdown==3.5.1
numpy==1.26.1